feat: Update GetAllCourierRequest and SearchCourier to use float64 for radius and enhance validation

+++ domain/courier/entity/courier.go
@@ -25,7 +25,7 @@ type SearchCourier struct {
 	Email     string
 	Longitude float64
 	Latitude  float64
-	Radius    int
+	Radius    float64
 	Page      int
 	PerPage   int
 	SortBy    string
